requirements.txt
requests>=2.28.0
pandas>=1.5.0
openpyxl>=3.0.0
python-dotenv>=1.0.0
PyQt6>=6.4.0
